requirements.txt
alembic==0.9.6
certifi==2017.11.5
chardet==3.0.4
click==6.7
coverage==4.4.2
coveralls==1.2.0
docopt==0.6.2
flake8==3.5.0
Flask==0.12.2
Flask-API==1.0
Flask-HTTPAuth==3.2.3
Flask-Migrate==2.1.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
nose==1.3.7
passlib==1.7.1
pew==1.1.0
pipenv==8.3.2
psutil==5.3.1
psycopg2==2.7.3.2
pycodestyle==2.3.1
pyflakes==1.6.0
PyJWT==1.5.3
python-dateutil==2.6.1
python-editor==1.0.3
requests==2.18.4
six==1.11.0
SQLAlchemy==1.1.15
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.2.6
Werkzeug==0.12.2
